+++ app/routes/users.php
@@ -12,6 +12,11 @@ $app->group('', function () use ($app) {
         // Change Password
         $this->map(['POST'], '/change-password', 'Profile:changePassword')
             ->setName('change-password');
+
+         $app->group('/2fa', function () use ($app) {
+            $this->post('[/{validate}]', 'Profile:twoFactor')
+                ->setName('2fa');
+         });
     });
 })
 ->add($container->get('csrf'))
@@ -19,13 +24,23 @@ $app->group('', function () use ($app) {
 ->add(new Dappur\Middleware\Maintenance($container))
 ->add(new Dappur\Middleware\PageConfig($container))
 ->add(new Dappur\Middleware\Seo($container))
-->add(new Dappur\Middleware\ProfileCheck($container));
+->add(new Dappur\Middleware\ProfileCheck($container))
+->add(new Dappur\Middleware\TwoFactorAuth($container));
 
 // Incomplete Profile Page
 $app->map(['GET','POST'], '/profile/incomplete', 'Profile:profileIncomplete')
-        ->setName('profile-incomplete')
-        ->add($container->get('csrf'))
-        ->add(new Dappur\Middleware\Auth($container))
-        ->add(new Dappur\Middleware\Maintenance($container))
-        ->add(new Dappur\Middleware\PageConfig($container))
-        ->add(new Dappur\Middleware\Seo($container));
+    ->setName('profile-incomplete')
+    ->add($container->get('csrf'))
+    ->add(new Dappur\Middleware\Auth($container))
+    ->add(new Dappur\Middleware\Maintenance($container))
+    ->add(new Dappur\Middleware\PageConfig($container))
+    ->add(new Dappur\Middleware\Seo($container))
+    ->add(new Dappur\Middleware\TwoFactorAuth($container));
+
+$app->map(['GET', 'POST'], '/2fa/confirm', 'Profile:twoFactorConfirm')
+    ->setName('2fa-confirm')
+    ->add($container->get('csrf'))
+    ->add(new Dappur\Middleware\Auth($container))
+    ->add(new Dappur\Middleware\Maintenance($container))
+    ->add(new Dappur\Middleware\PageConfig($container))
+    ->add(new Dappur\Middleware\Seo($container));